+++ examples/types.js
@@ -1,3 +1,14 @@
-import {createTypes} from 'ruto';
+function createTypes ({namespace, types}) {
+  if (!types) { throw new Error('types property required'); }
+  if (!Array.isArray(types)) { throw new Error('types must be an array'); }
+
+  const typesObject = {};
+
+  types.forEach(type => {
+    typesObject[type] = namespace ? `@@${namespace}/${type}` : type;
+  });
+
+  return typesObject;
+}
 
 export const types = createTypes({types: ['FETCH_EXAMPLE', 'FETCH_EXAMPLES']});