+++ src/ruto/index.js
@@ -1,6 +1,2 @@
-export {connect} from './location';
-export {createTypes} from './utilities';
-export {location} from './reducers';
 export {middleware} from './middleware';
 export {register} from './plugin';
-export {update} from './actions';