+++ src/ruto/types.js
@@ -1,3 +1,0 @@
-import {createTypes} from './utilities';
-
-export const types = createTypes({namespace: 'location', types: ['UPDATE']});